+++ lib/ts/recipe/emailpassword/components/themes/resetPasswordUsingToken/resetPasswordEmail.tsx
@@ -12,8 +12,7 @@
  * License for the specific language governing permissions and limitations
  * under the License.
  */
-import { useContext, useState } from "react";
-import StyleContext from "../../../../../styles/styleContext";
+import { useState } from "react";
 
 import { EnterEmailProps, EnterEmailStatus } from "../../../types";
 
@@ -27,7 +26,6 @@ import BackToSignInButton from "../../library/backToSignInButton";
 import BackButton from "../../library/backButton";
 
 const EmailPasswordResetPasswordEmail: React.FC<EnterEmailProps> = (props) => {
-    const styles = useContext(StyleContext);
     const t = useTranslation();
     const userContext = useUserContext();
     const [status, setStatus] = useState<EnterEmailStatus>("READY");
@@ -51,16 +49,11 @@ const EmailPasswordResetPasswordEmail: React.FC<EnterEmailProps> = (props) => {
 
     if (status === "SENT") {
         return (
-            <div data-supertokens="container" css={styles.container}>
-                <div data-supertokens="row" css={styles.row}>
-                    <div
-                        data-supertokens="primaryText enterEmailSuccessMessage"
-                        css={[styles.primaryText, styles.enterEmailSuccessMessage]}>
+            <div data-supertokens="container">
+                <div data-supertokens="row">
+                    <div data-supertokens="primaryText enterEmailSuccessMessage">
                         {emailSuccessText}
-                        <span
-                            data-supertokens="link resendEmailLink"
-                            css={[styles.link, styles.resendEmailLink]}
-                            onClick={resend}>
+                        <span data-supertokens="link resendEmailLink" onClick={resend}>
                             {t("EMAIL_PASSWORD_RESET_RESEND_LINK")}
                         </span>
                     </div>
@@ -72,23 +65,17 @@ const EmailPasswordResetPasswordEmail: React.FC<EnterEmailProps> = (props) => {
 
     // Otherwise, return Form.
     return (
-        <div data-supertokens="container" css={styles.container}>
-            <div data-supertokens="row" css={styles.row}>
-                <div
-                    data-supertokens="headerTitle resetPasswordHeaderTitle"
-                    css={[styles.headerTitle, styles.resetPasswordHeaderTitle]}>
+        <div data-supertokens="container">
+            <div data-supertokens="row">
+                <div data-supertokens="headerTitle resetPasswordHeaderTitle">
                     <BackButton onClick={props.onBackButtonClicked} />
                     {t("EMAIL_PASSWORD_RESET_HEADER_TITLE")}
-                    <span
-                        data-supertokens="backButtonPlaceholder backButtonCommon"
-                        css={[styles.backButtonPlaceholder, styles.backButtonCommon]}>
+                    <span data-supertokens="backButtonPlaceholder backButtonCommon">
                         {/* empty span for spacing the back button */}
                     </span>
                 </div>
-                <div data-supertokens="headerSubtitle" css={styles.headerSubtitle}>
-                    <div data-supertokens="secondaryText" css={styles.secondaryText}>
-                        {t("EMAIL_PASSWORD_RESET_HEADER_SUBTITLE")}
-                    </div>
+                <div data-supertokens="headerSubtitle">
+                    <div data-supertokens="secondaryText">{t("EMAIL_PASSWORD_RESET_HEADER_SUBTITLE")}</div>
                 </div>
                 {props.error !== undefined && <GeneralError error={props.error} />}
                 <FormBase