+++ lib/ts/recipe/thirdparty/components/features/signInAndUpCallback/index.tsx
@@ -20,9 +20,6 @@ import { Fragment, useCallback } from "react";
 import { Awaited, FeatureBaseProps } from "../../../../../types";
 import { useOnMountAPICall } from "../../../../../utils";
 import FeatureWrapper from "../../../../../components/featureWrapper";
-import { StyleProvider } from "../../../../../styles/styleContext";
-import { defaultPalette } from "../../../../../styles/styles";
-import { getStyles } from "../../themes/styles";
 import { CustomStateProperties } from "../../../types";
 import { SignInAndUpCallbackTheme } from "../../themes/signInAndUpCallback";
 import Recipe from "../../../recipe";
@@ -108,27 +105,18 @@ const SignInAndUpCallback: React.FC<PropType> = (props) => {
 
     const recipeComponentOverrides = useRecipeComponentOverrideContext();
 
-    const oAuthCallbackScreen = props.recipe.config.oAuthCallbackScreen;
-
     return (
         <ComponentOverrideContext.Provider value={recipeComponentOverrides}>
             <FeatureWrapper
                 useShadowDom={props.recipe.config.useShadowDom}
                 defaultStore={defaultTranslationsThirdParty}>
-                <StyleProvider
-                    rawPalette={props.recipe.config.palette}
-                    defaultPalette={defaultPalette}
-                    styleFromInit={oAuthCallbackScreen.style}
-                    rootStyleFromInit={props.recipe.config.rootStyle}
-                    getDefaultStyles={getStyles}>
-                    <Fragment>
-                        {/* No custom theme, use default. */}
-                        {props.children === undefined && <SignInAndUpCallbackTheme />}
+                <Fragment>
+                    {/* No custom theme, use default. */}
+                    {props.children === undefined && <SignInAndUpCallbackTheme config={props.recipe.config} />}
 
-                        {/* Otherwise, custom theme is provided, propagate props. */}
-                        {props.children}
-                    </Fragment>
-                </StyleProvider>
+                    {/* Otherwise, custom theme is provided, propagate props. */}
+                    {props.children}
+                </Fragment>
             </FeatureWrapper>
         </ComponentOverrideContext.Provider>
     );