+++ lib/ts/recipe/thirdpartyemailpassword/utils.ts
@@ -65,6 +65,6 @@ function normaliseSignInUpFeatureConfig(config?: SignInAndUpFeatureUserInput): N
         ...config,
         disableDefaultUI,
         defaultToSignUp,
-        style: config === undefined || config.style === undefined ? {} : config.style,
+        style: config === undefined || config.style === undefined ? "" : config.style,
     };
 }