+++ test/with-typescript/src/App.tsx
@@ -26,7 +26,6 @@ import Footer from "./Footer";
 import HeliumTheme from "./Themes/Helium";
 import HydrogenTheme from "./Themes/Hydrogen";
 import DarkTheme from "./Themes/Dark";
-import { CSSObject } from "@emotion/react";
 import Passwordless from "../../../recipe/passwordless";
 import { PasswordlessFlowType } from "supertokens-web-js/recipe/passwordless/types";
 import ThirdPartyPasswordless from "../../../recipe/thirdpartypasswordless";
@@ -207,15 +206,15 @@ function getQueryParams(param: string): string | null {
 
 export type Theme = {
     colors: Record<string, string>;
-    style: Record<string, CSSObject>;
+    style: Record<string, any>;
 };
 function getTheme(): {
     colors: Record<string, string>;
-    style?: Record<string, CSSObject>;
+    style?: string;
 } {
     let theme = {
         colors: {},
-        style: {},
+        style: "",
     };
 
     const themeParams = window.localStorage.getItem("useTheme");
@@ -226,11 +225,11 @@ function getTheme(): {
     }
 
     if (themeParams === "helium") {
-        return HeliumTheme;
+        // return HeliumTheme;
     }
 
     if (themeParams === "hydrogen") {
-        return HydrogenTheme;
+        // return HydrogenTheme;
     }
 
     return theme;
@@ -306,7 +305,6 @@ function getRecipeList() {
 
 function getEmailPasswordConfigs() {
     return EmailPassword.init({
-        palette: theme.colors,
         resetPasswordUsingTokenFeature: {
             enterEmailForm: {
                 style: theme.style,
@@ -400,7 +398,6 @@ function getThirdPartyConfigs() {
         async getRedirectionURL(context: ThirdPartyGetRedirectionURLContext) {
             return undefined;
         },
-        palette: theme.colors,
         signInAndUpFeature: {
             style: theme.style,
             privacyPolicyLink: "https://supertokens.io/legal/privacy-policy",
@@ -442,7 +439,6 @@ function getThirdPartyEmailPasswordConfigs() {
         async getRedirectionURL(context: ThirdPartyEmailPasswordGetRedirectionURLContext) {
             return undefined;
         },
-        palette: theme.colors,
         signInAndUpFeature: {
             style: theme.style,
             signUpForm: {